Create String Package
